Reorganised components & added Home Page Title
