+++ lib/ui/PageTitle.tsx
@@ -1,6 +1,6 @@
 type Props = {
   title: string;
-  description: string;
+  description?: string;
   image: string;
   alt: string;
 };
@@ -11,13 +11,13 @@ export default function PageTitle(props: Props) {
       <div className="flex flex-col md:flex-row items-center justify-center">
         <div className="w-3/4 md:w-1/3">
           <h1 className="mb-3 text-primary text-6xl">{props.title}</h1>
-          <p className="text-xl">{props.description}</p>
+          {props.description ? <p className="text-xl">{props.description}</p> : null}
         </div>
         <div className="hidden md:flex md:w-2/3">
           <img
             src={props.image}
             style={styles}
-            className="img-fluid mx-auto d-block rounded-md object-cover w-full h-72 lg:h-96"
+            className="img-fluid mx-auto d-block rounded-md object-cover w-full h-72 lg:h-96 my-2"
             alt={props.alt}
             loading="lazy"
           />
